course rollovers working, needs massive cleanup
